requirements.txt
easy-pil>=0.3.2
psutil>=5.9.8
python-dotenv>=1.0.1
pytube>=15.0.0
youtube-search>=2.1.2
yt-dlp>=2024.5.27
lyricsgenius>=3.0.1
nextcord>=2.6.0
nextcord[voice]>=2.6.0
nextcord[speed]>=2.6.0
termcolor>=2.4.0
PyYAML==6.0.1
mysql-connector-python>=8.3.0
xmltodict>=0.13.0
nest-asyncio>=1.6.0
meta-yt>=0.1.0b0